refactor(ai): relocate isEarlyColonialExpansion to observer_goal_phase (Refs #2509)
Move the early-colonial conquer-score bonus predicate out of the
legacy colonial_pressure module so it survives the planned S1
deletion of colonial_pressure.dart, mirroring PR #2746 which did
the same for hasColonialAcquisitionTargets. Production behaviour
is unchanged: the relocated body is field-equal to the legacy
compound (re-expressed as
hasColonialAcquisitionTargets(colonial) && newWorldProvincesOwned
< kColonialFewNwProvincesThreshold), and colonialBuildOrderThresholdCap
still inlines the acquisition-targets check so colonial_pressure.dart
does not import observer_goal_phase.dart and avoids a circular
library reference.

Imports: goal_manager.dart drops its now-unused colonial_pressure.dart
import — both predicates it consumed (hasColonialAcquisitionTargets,
isEarlyColonialExpansion) are now in observer_goal_phase.dart, which
the file already imports.

Tests: dedicated isEarlyColonialExpansion group moves to
observer_goal_phase_test.dart and grows to seven cases — three positive
arms (invadable NW, adjacent tribe owners, just-below-threshold
holdings), three negative arms (no acquisition targets, many NW
holdings, threshold boundary), and a determinism check (Must-have #7).
colonial_pressure_test.dart drops the relocated symbol's import-show
and the boundary group, leaving its preamble pointing at the new home.

SPEC: phase-planner-architecture.md § Phase transition guard notes
that the goal-scoring sibling lives alongside the COLONIAL guard for
the same retirement-survival reason.

Refs #2509 — does not auto-close.
