+++ packages/colonizethis_ai/lib/src/planning/goal_manager.dart
@@ -4,7 +4,6 @@ import 'planning_imports.dart';
 
 import '../util/ai_random_utils.dart';
 import '../perception/perception_snapshot.dart';
-import 'colonial_pressure.dart';
 import 'observer_goal_phase.dart';
 import 'phase_planner_goal_filter.dart';
